chore: addressing pr feedback
