+++ src/.defaultsForValidator.js
@@ -35,8 +35,10 @@ const defaults = {
       'content_type_parameter': 'error',
       'accept_type_parameter': 'error',
       'authorization_parameter': 'warning',
-      'required_param_has_default': 'warning',
-      'invalid_limit_type': 'warning'
+      'required_param_has_default': 'warning'
+    },
+    'pagination_parameters': {
+      'pagination_style': 'warning'
     },
     'paths': {
       'missing_path_parameter': 'error',
