+++ src/plugins/validation/2and3/semantic-validators/parameters-ibm.js
@@ -26,99 +26,6 @@ module.exports.validate = function({ jsSpec, isOAS3 }, config) {
       return;
     }
     const contentsOfParameterObject = isParameter(path, isOAS3);
-    for (const head in obj.paths) {
-      let topLevelIsArray = false;
-      //loop through responses and see if there is an array in the top level
-      if (obj.paths[head].get) {
-        for (const prop in obj.paths[head].get.responses) {
-          if (obj.paths[head].get.responses[prop]) {
-            for (const mediaTypeKey in obj.paths[head].get.responses[prop]
-              .content) {
-              for (const param in obj.paths[head].get.responses[prop].content[
-                mediaTypeKey
-              ].schema) {
-                if (
-                  Array.isArray(
-                    obj.paths[head].get.responses[prop].content[mediaTypeKey]
-                      .schema[param]
-                  )
-                ) {
-                  topLevelIsArray = true;
-                  break;
-                }
-              }
-            }
-          }
-        }
-        if (obj.paths[head].get && topLevelIsArray) {
-          if (
-            obj.paths[head].get.parameters[0].limit &&
-            (typeof obj.paths['/pets'].get.parameters[0].limit !== 'number' ||
-              obj.paths[head].get.parameters[0].required.includes('limit'))
-          ) {
-            const message =
-              'limit parameter is optional and must be of type integer and must have a default value';
-            const checkStatus = config.invalid_limit_type;
-            if (checkStatus !== 'off') {
-              result[checkStatus].push({
-                path,
-                message
-              });
-            }
-          }
-          if (
-            obj.paths[head].get.parameters[0].start &&
-            (typeof obj.paths[head].get.parameters[0].start !== 'string' ||
-              obj.paths[head].get.parameters[0].required == true)
-          ) {
-            const message =
-              'start parameter must be of type string and must be optional';
-            const checkStatus = config.invalid_limit_type;
-            if (checkStatus !== 'off') {
-              result[checkStatus].push({
-                path,
-                message
-              });
-            }
-          }
-          if (
-            obj.paths[head].get.parameters[0].cursor &&
-            (typeof obj.paths[head].get.parameters[0].cursor !== 'string' ||
-              obj.paths[head].get.parameters[0].required == true)
-          ) {
-            const message =
-              'cursor parameter must be of type string and must be optional';
-            const checkStatus = config.invalid_limit_type;
-            if (checkStatus !== 'off') {
-              result[checkStatus].push({
-                path,
-                message
-              });
-            }
-          }
-          if (
-            (!obj.paths['/pets'].get.parameters[0].start ||
-              !obj.paths['/pets'].get.parameters[0].cursor) &&
-            obj.paths['/pets'].get.parameters[0].offset &&
-            (typeof obj.paths['/pets'].get.parameters[0].offset !== 'number' ||
-              obj.paths['/pets'].get.parameters[0].required.includes(
-                'offset'
-              ) == true)
-          ) {
-            const message =
-              'if start or cursor parameters are not present then the offset parameter should be defined as an integer and should be optional';
-            const checkStatus = config.invalid_limit_type;
-            if (checkStatus !== 'off') {
-              result[checkStatus].push({
-                path,
-                message
-              });
-            }
-          }
-        }
-      }
-    }
-
     if (contentsOfParameterObject) {
       // obj is a parameter object
       const isRef = !!obj.$ref;
